Add NCNN and VGGFace examples for ACE concept discovery pipeline
